refactor: extract DTOs to separate dto.rs module
Move all Tauri DTOs and From trait implementations from main.rs
to a dedicated dto.rs module for better code organization.

- Create dto.rs with all DTO structs organized by category
- Add module documentation explaining the purpose
- Update main.rs to import from dto module
- Reduce main.rs from ~770 lines to ~450 lines
